backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
apscheduler==3.10.4
requests==2.31.0
kiteconnect==5.0.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pytz==2024.1
